Move gemoji dagger module to gemoji gradle module
