+++ gemoji/build.gradle
@@ -61,10 +61,14 @@ kapt {
 }
 
 dependencies {
-  kapt deps.android.arch.room.apt
+  implementation project(':util')
 
-  api deps.kotlin.stdlib.jdk7
+  kapt deps.android.arch.room.apt
+  kapt deps.dagger.apt.compiler
+
   api deps.android.arch.room.runtime
+  api deps.dagger.runtime
+  api deps.kotlin.stdlib.jdk7
   api deps.misc.moshi
 
   testImplementation deps.test.junit
