+++ gemoji/src/main/kotlin/io/sweers/catchup/gemoji/GemojiDatabase.kt
@@ -16,21 +16,10 @@
 
 package io.sweers.catchup.gemoji
 
-import android.arch.persistence.db.framework.SupportAssetSQLiteOpenHelper
 import android.arch.persistence.room.Database
-import android.arch.persistence.room.Room
 import android.arch.persistence.room.RoomDatabase
-import android.content.Context
 
-fun createGemojiDatabase(context: Context, name: String): GemojiDatabase {
-  return Room.databaseBuilder(context, GemojiDatabase::class.java, name)
-      .openHelperFactory {
-        SupportAssetSQLiteOpenHelper(it.context, it.name!!, it.callback.version, it.callback)
-      }
-      .build()
-}
-
-@Database(entities = arrayOf(Gemoji::class), version = 1)
+@Database(entities = [(Gemoji::class)], version = 1)
 abstract class GemojiDatabase: RoomDatabase() {
   abstract fun GemojiDao(): GemojiDao
 }
